Going to change how the concurrency works.
Instead of doing it in batches of core_num (letting each core do one frame), I will change it so when a frame is being rendered, I will have a thread pool and each worker will be given one area of the frame to render, returning the computed area. May be better for renders with higher iterations and deeper zooms.